+++ src/pytest_embedded_arduino_cli/__init__.py
@@ -2,5 +2,5 @@
 
 __all__ = ["__version__"]
 
-__version__ = "1.1.7"
+__version__ = "1.2.0"